Move core channel state into coach - the only place it is still used.
